Redirecionando para a pagina home, quando houver atualização da pagina do reositórios

+++ src/pages/Repositories/index.js
@@ -1,17 +1,23 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styled';
+import { useHistory } from 'react-router-dom';
 
 
 
 export default function Repositories() {
+    const history = useHistory();
     const [ repositories, setRepoitories ] = useState([]);
 
 
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepoitories(repositoriesName);
-        localStorage.clear();
+        if(repositoriesName !== null){
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepoitories(repositoriesName);
+            localStorage.clear();
+        }else {
+            history.push('/');
+        }        
     }, []);
 
     return (
